put alpha_max in new file
